requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
soupsieve==2.5
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
